More generic setup for final container image.
Final container needs python and python-venv installed, but the version
that should be used is dictated by the install script, with no simple
way to just extract the portion of the script that decides that.

Instead, look at the python version that the installed venv (from the
build container) is using, and install that and distutils. This removes
the dependency on the information contained in the install script, but
assumes the directory layout for the venv won't change and the ubuntu
python packages always follow a specific naming scheme.