Add possibility to set multiple services in "service" environment variable (#150)
This change will allow "service" variable to take multiple options and function
in the same way as "chia start" command. For example, it will be possible
to start container with -e service="farmer-only harvester wallet" to start only
chosen services - farmer, harvester and wallet in this case.
This was previously not possible without entrypoint overriding.